requirements.txt
python-dotenv==1.1.0
moviepy==2.2.1
openai==1.86.0
